minor refinements

+++ src/queue.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 
 // Add the item to the end
-void queue_push(struct RecordQueue* q, const char* destination, const char* departure_datetime, const char* arrival_datetime, const char* type_of_coach, double ticket_price, const char* purchase_datetime, bool available) {
+void queue_push(struct RecordQueue* q, const char* destination, const char* departure_datetime, const char* arrival_datetime, const char* type_of_coach, float ticket_price, const char* purchase_datetime, bool available) {
 	if (!q->head) {
 		q->head = q->tail = record_new(q->nextID, destination, departure_datetime, arrival_datetime, type_of_coach, ticket_price, purchase_datetime, available);
 		q->nextID++;
